+++ base.ts
@@ -189,6 +189,7 @@ export namespace LintError {
 	export interface Fix {
 		readonly range: [number, number];
 		text: string;
+		desc: string;
 	}
 }