+++ src/attribute.ts
@@ -211,14 +211,9 @@ export abstract class AttributeToken extends Token {
 			);
 			e.startIndex--;
 			e.startCol--;
-			const fix: LintError.Fix = {range: [e.endIndex, e.endIndex], text: this.#quotes[0]!};
+			const fix: LintError.Fix = {range: [e.endIndex, e.endIndex], text: this.#quotes[0]!, desc: 'close'};
 			if (lastChild.childNodes.some(({type: t, data}) => t === 'text' && /\s/u.test(data))) {
-				e.suggestions = [
-					{
-						desc: 'close',
-						...fix,
-					},
-				];
+				e.suggestions = [fix];
 			} else {
 				e.fix = fix;
 			}
@@ -322,9 +317,9 @@ export abstract class AttributeToken extends Token {
 			{childNodes} = Parser.parse(value, this.getAttribute('include'), stages[this.type] + 1, config);
 		this.lastChild.replaceChildren(...childNodes);
 		if (value.includes('"')) {
-			this.#quotes = [`'`, `'`] as [string, string];
+			this.#quotes = [`'`, `'`] as const;
 		} else if (value.includes(`'`) || !this.#quotes[0]) {
-			this.#quotes = ['"', '"'] as [string, string];
+			this.#quotes = ['"', '"'] as const;
 		} else {
 			this.close();
 		}