+++ src/link/base.ts
@@ -238,6 +238,7 @@ export abstract class LinkBaseToken extends Token {
 				e.fix = {
 					range: [e.startIndex + textNode.getRelativeIndex() + textNode.data.indexOf('#'), e.endIndex],
 					text: '',
+					desc: 'remove',
 				};
 			}
 			errors.push(e);